+++ internal/testutils/mock/mines.sql
@@ -2,7 +2,8 @@ DROP TABLE IF EXISTS mines;
 CREATE TABLE mines (
                        id    INTEGER PRIMARY KEY,
                        mine_name TEXT  DEFAULT '',
-                       purl_type  TEXT  DEFAULT ''
+                       purl_type  TEXT  DEFAULT '',
+                       repository_url TEXT DEFAULT ''
 );
 INSERT INTO mines (id, mine_name, purl_type) VALUES (0, 'maven.org', 'maven');
 INSERT INTO mines (id, mine_name, purl_type) VALUES (1, 'rubygems.org', 'gem');